BE_py/book-platform/requirements.txt

# Core FastAPI
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
python-dotenv>=1.0.0
python-multipart>=0.0.9

# Database
mysql-connector-python>=9.0.0

httpx>=0.27.0

# Search (OpenSearch)
opensearch-py>=2.5.0

scikit-learn>=1.5.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0

# Chatbot – Embedding & Vector Search
sentence-transformers>=3.0.0
faiss-cpu>=1.8.0

underthesea>=6.8.4

httpx>=0.27.0
